Refactor/UI package (#4)
* refactor the utilities package

* use env file

* update pre-commit to hook to pull from new repo dev branch

* public env file

* rename package from control to main

* refactor pages and components

* more refactoring to the control package; only the sidebar components is remaining for refactoring

* refactor the sidebar component and fix bug in the topbar component

* implement changes in previous conversation on the refactor/zuri_ui branch
