--- packages/ui/src/utils/index.js
@@ -1,2 +1,0 @@
-export * from "./get-sample-messages";
-export * from "./get-sample-member-list";
